# docs/contracts/admin.openapi.yaml
components:
  schemas:
...
    DateTime:
      description: ISO-8601; Zod wire type z.iso.datetime()
      format: date-time
      pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
      type: string
...
    ErrorEnvelope:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/__schema29"
      required:
        - error
      type: object
...
    Id:
      format: uuid
      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
      type: string
    OperatorTokenCreate:
      properties:
        label:
          $ref: "#/components/schemas/__schema19"
      required:
        - label
      type: object
    OperatorTokenCreated:
      additionalProperties: false
      properties:
        createdAt:
          $ref: "#/components/schemas/DateTime"
        id:
          $ref: "#/components/schemas/Id"
        label:
          $ref: "#/components/schemas/__schema44"
        token:
          $ref: "#/components/schemas/__schema48"
      required:
        - id
        - token
        - label
        - createdAt
      type: object
    OperatorTokenList:
      additionalProperties: false
      properties:
        items:
          $ref: "#/components/schemas/__schema43"
        nextCursor:
          $ref: "#/components/schemas/__schema47"
      required:
        - items
        - nextCursor
      type: object
    OperatorTokenSummary:
      additionalProperties: false
      properties:
        createdAt:
          $ref: "#/components/schemas/DateTime"
        id:
          $ref: "#/components/schemas/Id"
        label:
          $ref: "#/components/schemas/__schema44"
        lastUsedAt:
          $ref: "#/components/schemas/__schema46"
        prefix:
          $ref: "#/components/schemas/__schema45"
      required:
        - id
        - label
        - prefix
        - createdAt
        - lastUsedAt
      type: object
...
    __schema19:
      description: Operator-chosen name identifying the holder, e.g. a CI job or laptop
      maxLength: 200
      minLength: 1
      type: string
...
    __schema29:
      additionalProperties: false
      properties:
        code:
          description: Stable machine code (e.g. validation_failed, not_found, unauthorized, conflict)
          type: string
        details:
          $ref: "#/components/schemas/__schema30"
          description: Present for validation_failed (Zod issues)
        message:
          type: string
      required:
        - code
        - message
      type: object
...
    __schema30:
      items:
        additionalProperties: false
        properties:
          message:
            type: string
          path:
            type: string
        required:
          - path
          - message
        type: object
      type: array
...
    __schema43:
      items:
        $ref: "#/components/schemas/OperatorTokenSummary"
      type: array
    __schema44:
      description: Operator-chosen name identifying the holder, e.g. a CI job or laptop
      maxLength: 200
      minLength: 1
      type: string
    __schema45:
      description: First ~8 chars of the token for recognition; never full secret
      type: string
    __schema46:
      anyOf:
        - $ref: "#/components/schemas/DateTime"
        - type: "null"
    __schema47:
      anyOf:
        - type: string
        - type: "null"
    __schema48:
      description: Plaintext operator token — shown once
      type: string
...
  securitySchemes:
    OperatorBearer:
      bearerFormat: API key
      description: "The bootstrap OPERATOR_API_KEY, or any minted operator_token (issue #41) — both are accepted as Bearer credentials with equal privilege. Only the bootstrap key may also be exchanged for the dashboard's HttpOnly session cookie via POST /auth/login; the cookie itself is accepted here as an alternate to Bearer, not a separate OAuth/user session."
      scheme: bearer
      type: http
    OperatorCookie:
      description: Optional alternate to OperatorBearer for the dashboard.
      in: cookie
      name: wb_operator
      type: apiKey
info:
  description: Operator-facing admin contract emitted from Zod schemas in packages/contract. Ingest `POST /ingest/{slug}` is intentionally NOT in this document.
  title: webhook-broadcast Admin API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /operator-tokens:
    get:
      operationId: listOperatorTokens
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorTokenList"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - OperatorBearer: []
      summary: "List operator token summaries (issue #41)"
      tags:
        - operator-tokens
    post:
      operationId: createOperatorToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorTokenCreate"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorTokenCreated"
          description: Created
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - OperatorBearer: []
      summary: Mint an operator token; plaintext returned once
      tags:
        - operator-tokens
  /operator-tokens/{tokenId}:
    delete:
      operationId: revokeOperatorToken
      parameters:
        - in: path
          name: tokenId
          required: true
          schema:
            $ref: "#/components/schemas/Id"
      responses:
        "204":
          description: Revoked
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error
      security:
        - OperatorBearer: []
      summary: Revoke an operator token immediately, no redeploy required
      tags:
        - operator-tokens
security:
  - OperatorBearer: []
servers:
  - url: /
...